requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
atlassian-python-api==3.41.13
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cssselect2==0.7.0
dataclasses-json==0.6.6
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.1
faiss-cpu==1.8.0
freetype-py==2.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.16
langsmith==0.1.63
lxml==5.2.2
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.3
orjson==3.10.3
packaging==23.2
pdf2image==1.17.0
pillow==10.3.0
pycairo==1.26.0
pydantic==2.7.1
pydantic_core==2.18.2
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
replicate==0.26.0
reportlab==4.2.0
requests==2.32.2
requests-oauthlib==2.0.0
rlPyCairo==0.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
svglib==1.5.1
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
tqdm==4.66.4
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
webencodings==0.5.1
wrapt==1.16.0
yarl==1.9.4
